% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_heat_map.R
\name{DE_heat_map}
\alias{DE_heat_map}
\title{Creates a heat map}
\usage{
DE_heat_map(
  count_df,
  col_data,
  dds,
  res,
  xvar,
  name_of_comparison,
  top_n = 50,
  genes_of_interest = NULL
)
}
\arguments{
\item{count_df}{RNA Seq raw counts dataframe, including columns gene_id and gene_name}

\item{col_data}{Column data (sample metadata)}

\item{dds}{DESeq dataset}

\item{res}{DESeq2 results (after running results() on DESeq dataset)}

\item{xvar}{Coefficient or design of experiment}

\item{name_of_comparison}{Name of comparison}

\item{top_n}{Number of genes to plot (default 50)}

\item{genes_of_interest}{List of genes of interest (default none)}
}
\value{
A heat map
}
\description{
Creates a heat map
}
