% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_pca.R
\name{DE_pca}
\alias{DE_pca}
\title{Creates a Principal Component Analysis (PCA) plot}
\usage{
DE_pca(dds, xvar, top_n = 500, PC_of_interest_x = 1, PC_of_interest_y = 2)
}
\arguments{
\item{dds}{DESeq dataset}

\item{xvar}{Coefficient or design of experiment (column in metadata)}

\item{top_n}{Number of top features to use (default 500)}

\item{PC_of_interest_x}{PC to be on x axis (default 1)}

\item{PC_of_interest_y}{PC to be on y axis (detaulf 2)}
}
\value{
A PCA
}
\description{
Creates a Principal Component Analysis (PCA) plot
}
