% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_table_volcano.R
\name{DE_table_volcano}
\alias{DE_table_volcano}
\title{Creates a volcano plot}
\usage{
DE_table_volcano(
  count_df,
  dds,
  res,
  shrinkage_method,
  xvar,
  name_of_comparison,
  first_group,
  second_group
)
}
\arguments{
\item{count_df}{RNA Seq raw counts dataframe, including columns gene_id and gene_name}

\item{dds}{DESeq dataset}

\item{res}{DESeq2 results (after running results() on DESeq dataset)}

\item{shrinkage_method}{Shrinkage method (apeglm, ashr, or normal)}

\item{xvar}{Coefficient or design of experiment}

\item{name_of_comparison}{Name of comparison (column in metadata)}

\item{first_group}{Name of first group of comparison}

\item{second_group}{Name of second group of comparison}
}
\value{
A volcano plot
}
\description{
Creates a volcano plot
}
