% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_ranking_and_gmt.R
\name{prepare_gmt}
\alias{prepare_gmt}
\title{Prepares gmt files to be used in fgsea}
\usage{
prepare_gmt(gmt_file, genes_in_data, savefile = FALSE)
}
\arguments{
\item{gmt_file}{gmt files containing genes sets}

\item{genes_in_data}{A list of genes present in the data}
}
\value{
A prepared gmt file
}
\description{
Prepares gmt files to be used in fgsea
}
