% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_ranking_and_gmt.R
\name{prepare_rankings}
\alias{prepare_rankings}
\title{Prepares a ranked list of genes using a signed p-value metric to be used in fgsea}
\usage{
prepare_rankings(log2foldchange, pvalues, row_names)
}
\arguments{
\item{log2foldchange}{Log2 fold change of genes}

\item{pvalues}{P values of genes}

\item{row_names}{Gene names}
}
\value{
A ranked list of genes
}
\description{
Prepares a ranked list of genes using a signed p-value metric to be used in fgsea
}
